ipsec-tools: Do not run the phase1-up script more than once.
Racoon executes phase1-up or phase1-down when phase 1 is established or
destroyed. Combining with the usage of mode-cfg, phase1-up generates the
SPD policy and sets the internal network, and phase1-down reverts what
phase1-up does. However, a new phase 1 may be established before the
previous one is destroyed, so phase1-down may be executed after the new
phase1-up, which causes some difficulties to maintain the state.

It can be solved in several ways, such as disabling phase 1 rekeying or
carefully updating the internal network. The former limits the lifetime
of a VPN connection, and the latter complicates the VPN on Android. We
can also disable mode-cfg or skip the phase1-up script after the first
phase 1 is established. It is hard to tell which one is the best. Just
try one and see if we need to switch to another.

Change-Id: Icd010c3242b2f66651e1332f819fb5f274c9fc45