ipsec-tools: update for BoringSSL.
BoringSSL is Google's version of OpenSSL[1] with lots of things cleaned
up and removed. This change makes ipsec-tools work with both BoringSSL
and Android's existing version of OpenSSL.

Some changes are generally applicable but upstream ipsec-tools appears
to be dead. I've followed the existing practice of #ifdef'ing the
changes but, if there's no upstream, then I'd be happy to be more
aggressive about that.

Otherwise, apart from a couple of tweaks, the big changes are to the way
that the ENGINE works and to replace the PKCS#7 parsing code.

[1] https://www.imperialviolet.org/2014/06/20/boringssl.html

Change-Id: I32d7eb4e0e0a90f92cf5e8f9675dac4a8edc7a4a